Be/dev 병합

See merge request s11-bigdata-recom-sub1/S11P21A509!148

+++ src/main/java/ssafy/ddada/common/util/RatingUtil.java
@@ -23,7 +23,7 @@ public class RatingUtil {
         return (int) Math.round(team.stream().mapToInt(Player::getRating).average().orElse(0));
     }
 
-    public Integer updatePlayerRating(Player player, double opponentTeamRating, boolean isWin, Integer teamScore, int earnedRate, int missedRate) {
+    public Integer updatePlayerRating(Player player, double opponentTeamRating, boolean isWin, Integer teamScore, double earnedRate, double missedRate) {
         int minChange = -100;
         int maxChange = 100;
